requirements.txt
pandas
numpy
requests
google-cloud-translate

